requirements.txt
PyMuPDF
Pillow
pytesseract
requests
pytest

